# docs/swagger.yaml
definitions:
...
  requestmodel.AdminLoginData:
    properties:
      email:
        type: string
      password:
        minLength: 4
        type: string
    type: object
...
  requestmodel.Cart:
    properties:
      cartid_userid:
        type: string
      price:
        type: integer
      productid:
        type: string
      quantity:
        maximum: 5
        minimum: 1
        type: integer
    required:
    - productid
    - quantity
    type: object
...
  response.Response:
    properties:
      error: {}
      message:
        type: string
      result: {}
      status_code:
        type: integer
    type: object
...
info:
  contact:
    email: shahabazsulthan4@gmail.com
    name: 'Creator: Shahabaz Sulthan'
  description: "Laptop Lounge - \U0001F680 Your One-Stop Destination for Ultimate
    Laptop Shopping!\n\U0001F6D2\U0001F4BB Browse, compare, and buy top-notch laptops
    effortlessly.\nPowered by cutting-edge technology, we bring you a seamless shopping
    experience.\n\U0001F31F Dive into the future of laptop shopping with Laptop Lounge!
    \U0001F680\U0001F525"
  title: Laptop Lounge API
paths:
...
  /admin/:
    get:
      consumes:
      - application/json
      description: Retrieve details for the admin. Requires a valid Bearer token.
      produces:
      - application/json
      responses:
        "200":
          description: Admin details retrieved successfully
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized. Authentication required.
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error.
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Get Admin Dashboard Details
      tags:
      - Admins
...
  /admin/login:
    post:
      consumes:
      - application/json
      description: Using this handler, admins can log in and receive an authentication
        token.
      parameters:
      - description: Admin login details
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/requestmodel.AdminLoginData'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully logged in. Token returned.
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request. Invalid input.
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized. Authentication failed.
          schema:
            $ref: '#/definitions/response.Response'
      summary: Admin Login
      tags:
      - Admins
...
  /cart/:
    get:
      consumes:
      - application/json
      description: Retrieve all items in the user's cart.
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved user cart items
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      - RefreshtokenAuth: []
      summary: Get User Cart
      tags:
      - UserCart
    post:
      consumes:
      - application/json
      description: Create a user cart.
      parameters:
      - description: Cart details for creating
        in: body
        name: cart
        required: true
        schema:
          $ref: '#/definitions/requestmodel.Cart'
      produces:
      - application/json
      responses:
        "200":
          description: User cart created successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      - RefreshtokenAuth: []
      summary: Create User Cart
      tags:
      - UserCart
  /cart/{productID}:
    delete:
      consumes:
      - application/json
      description: Delete a product from the user's cart.
      parameters:
      - description: Product ID to delete from the cart
        in: path
        name: productID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Product deleted from the cart successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      - Refreshtoken: []
      summary: Delete Item from User Cart
      tags:
      - UserCart
...
securityDefinitions:
  BearerTokenAuth:
    description: 'Use your Bearer token for authentication. Example: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
  RefreshTokenAuth:
    description: 'Use your Refresh token to obtain a new Bearer token. Example: "{token}"'
    in: header
    name: Refreshtoken
    type: apiKey
swagger: "2.0"